Sheep model has been properly swapped out but I broke Wander()
