+++ Assets/Blends.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 59e1b0743b4714e059e40ec8c3ed81fc
+folderAsset: yes
+timeCreated: 1516758837
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
