+++ Assets/Objects/Raptor/Animations.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 68b2d83a3d9ef46298f8ceeaf4515b1e
+folderAsset: yes
+timeCreated: 1516758225
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
