+++ Assets/Objects/Raptor/Textures/raptor.jpg.meta
