+++ Assets/Objects/Raptor/Textures/raptor_normal.jpg
