+++ Assets/Objects/RaptorPenHouse/Prefabs.meta
@@ -1,9 +1,10 @@
 fileFormatVersion: 2
-guid: 96419a142fcfd43bc8eb8db03b16707e
+guid: ea9ce9d1303d349908068f059df3d709
 folderAsset: yes
-timeCreated: 1512957904
+timeCreated: 1516759369
 licenseType: Free
 DefaultImporter:
+  externalObjects: {}
   userData: 
   assetBundleName: 
   assetBundleVariant:
