+++ Assets/Objects/Sheep/Prefabs/Sheep.prefab.meta
@@ -1,9 +1,10 @@
 fileFormatVersion: 2
-guid: bd2e115e420022544848e525ef9b6bf4
-timeCreated: 1510787720
+guid: 57b4d44a3feed4a828ef15a5bb7d22ff
+timeCreated: 1516757872
 licenseType: Free
 NativeFormatImporter:
-  mainObjectFileID: 11400000
+  externalObjects: {}
+  mainObjectFileID: 100100000
   userData: 
   assetBundleName: 
   assetBundleVariant:
