+++ Assets/Objects/Sheep/Textures.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 89affe865687a48aea363c2d113eb5e7
+folderAsset: yes
+timeCreated: 1516758156
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
