+++ Assets/Objects/Sheep/Textures/SheepTex.png
