+++ Assets/Objects/Sheep/Textures/SheepUV.png.meta
