--- Assets/Objects/Sheepy/CubeSheepy.controller.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 07d399c422c141b4fbaf33d08346f415
-timeCreated: 1507920294
-licenseType: Free
-NativeFormatImporter:
-  mainObjectFileID: 9100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
