--- Assets/Objects/Sheepy/Graze.anim.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 42e9c7121169ad541b1239131414990a
-timeCreated: 1507920293
-licenseType: Free
-NativeFormatImporter:
-  mainObjectFileID: 7400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
