+++ Assets/Scripts/Sheeps/Basic Sheep.asset
@@ -8,7 +8,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 0cf2f7e9eb53f39449de40accb0e0f4a, type: 3}
+  m_Script: {fileID: 11500000, guid: 817f2e5afa1bc46759cc5ea76f657971, type: 3}
   m_Name: Basic Sheep
   m_EditorClassIdentifier: 
   healthStarting: 15
@@ -22,5 +22,5 @@ MonoBehaviour:
   petrifiedDistance: 30
   moveSpeed: 10
   wanderDistance: 35
-  rotationSpeed: 100
+  rotationSpeed: 5
   angleOfSight: 25
