+++ Assets/Scripts/Sheeps/Basic Sheep.asset.meta
@@ -1,9 +1,10 @@
 fileFormatVersion: 2
-guid: 0fcc7a5dea5254510bcf947823e9025f
-timeCreated: 1512957961
+guid: 8c0735f8fa3f04a1d9e5950b0041dec7
+timeCreated: 1516760720
 licenseType: Free
 NativeFormatImporter:
-  mainObjectFileID: 100100000
+  externalObjects: {}
+  mainObjectFileID: 11400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
