+++ Assets/Scripts/Sheeps/SheepEnemy.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 0cf2f7e9eb53f39449de40accb0e0f4a
-timeCreated: 1510615015
+guid: 26418f60ef80c1e4f961a5da3db535ce
+timeCreated: 1510615436
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
