+++ Assets/Scripts/Sheeps/SheepEnemyType.cs.meta
@@ -1,8 +1,9 @@
 fileFormatVersion: 2
-guid: 26418f60ef80c1e4f961a5da3db535ce
-timeCreated: 1510615436
+guid: 817f2e5afa1bc46759cc5ea76f657971
+timeCreated: 1516760278
 licenseType: Free
 MonoImporter:
+  externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
